Implemented Strategy Design Pattern in Spring using annotation
While implementing strategy Design Pattern in spring using annotations, We don't need to implement factory class or factory method.

Because, IOC container takes the responsibility of creating the objects/beans for target and dependent class and Injecting dependent beans to target class. So it reduces the burden on developer of creating extra class for implementing factory method.